+++ internal/chunk/sample_readjob.go
@@ -25,7 +25,7 @@ type (
 
 	ReadJobResult struct {
 		Err           error
-		Chunk         SampleChunk
+		Chunk         *Chunk
 		TransactionID string
 		ThreadID      *string
 		Start         uint64
@@ -34,7 +34,7 @@ type (
 )
 
 func (job ReadJob) Read() {
-	var chunk SampleChunk
+	var chunk Chunk
 
 	err := storageutil.UnmarshalCompressed(
 		job.Ctx,
@@ -45,7 +45,7 @@ func (job ReadJob) Read() {
 
 	job.Result <- ReadJobResult{
 		Err:           err,
-		Chunk:         chunk,
+		Chunk:         &chunk,
 		TransactionID: job.TransactionID,
 		ThreadID:      job.ThreadID,
 		Start:         job.Start,
@@ -63,7 +63,7 @@ type (
 	CallTreesReadJobResult struct {
 		Err           error
 		CallTrees     map[string][]*nodetree.Node
-		Chunk         SampleChunk
+		Chunk         *SampleChunk
 		TransactionID string
 		ThreadID      *string
 		Start         uint64
@@ -90,7 +90,7 @@ func (job CallTreesReadJob) Read() {
 	job.Result <- CallTreesReadJobResult{
 		Err:           err,
 		CallTrees:     callTrees,
-		Chunk:         chunk,
+		Chunk:         &chunk,
 		TransactionID: job.TransactionID,
 		ThreadID:      job.ThreadID,
 		Start:         job.Start,