+++ internal/profile/android.go
@@ -1,6 +1,7 @@
 package profile
 
 import (
+	"errors"
 	"fmt"
 	"hash/fnv"
 	"math"
@@ -242,6 +243,62 @@ func (p *Android) FixSamplesTime() {
 			p.Events[i].Time.Monotonic.Wall.Nanos = (newTime % 1e9)
 		}
 	}
+}
+
+func (p *Android) AddTimeDelta(deltaNS int64) func(*AndroidEvent) error {
+	var addDeltaTimestamp func(e *AndroidEvent) error
+	timestampBuilder := p.TimestampGetter()
+	switch p.Clock {
+	case GlobalClock:
+		addDeltaTimestamp = func(e *AndroidEvent) error {
+			ts := timestampBuilder(e.Time)
+			ts, err := getTsFromDelta(ts, deltaNS)
+			if err != nil {
+				return err
+			}
+			secs := (ts / 1e9)
+			nanos := (ts % 1e9)
+			e.Time.Global.Secs = secs
+			e.Time.Global.Nanos = nanos
+			return nil
+		}
+	case CPUClock:
+		addDeltaTimestamp = func(e *AndroidEvent) error {
+			ts := timestampBuilder(e.Time)
+			ts, err := getTsFromDelta(ts, deltaNS)
+			if err != nil {
+				return err
+			}
+			secs := (ts / 1e9)
+			nanos := (ts % 1e9)
+			e.Time.Monotonic.CPU.Secs = secs
+			e.Time.Monotonic.CPU.Nanos = nanos
+			return nil
+		}
+	default:
+		addDeltaTimestamp = func(e *AndroidEvent) error {
+			ts := timestampBuilder(e.Time)
+			ts, err := getTsFromDelta(ts, deltaNS)
+			if err != nil {
+				return err
+			}
+			secs := (ts / 1e9)
+			nanos := (ts % 1e9)
+			e.Time.Monotonic.Wall.Secs = secs
+			e.Time.Monotonic.Wall.Nanos = nanos
+			return nil
+		}
+	}
+	return addDeltaTimestamp
+}
+
+func getTsFromDelta(ts uint64, deltaNS int64) (uint64, error) {
+	if deltaNS < 0 && uint64(-deltaNS) <= ts {
+		return ts - uint64(-deltaNS), nil
+	} else if deltaNS >= 0 {
+		return ts + uint64(deltaNS), nil
+	}
+	return 0, errors.New("error: cannot subtract a delta bigger than the timestamp itself")
 }
 
 // CallTrees generates call trees for a given profile.